requirements.txt
# coverage==4.5.1
Django==2.0.5
# django-debug-toolbar==1.9.1
django-filter==1.1.0
djangorestframework==3.8.2
Markdown==2.6.11
Pillow==5.1.0
pytz==2018.4
sqlparse==0.2.4
